requirements.txt
azure-functions==1.17.0
cmdstanpy==1.2.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
fonttools==4.46.0
holidays==0.38
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
plotly==5.18.0
prophet==1.1.5
psycopg2-binary==2.9.9
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
scipy==1.11.4
six==1.16.0
stanio==0.3.0
tenacity==8.2.3
tqdm==4.66.1
tzdata==2023.3
xgboost==2.0.2
xlrd==2.0.1
zipp==3.17.0
scikit-learn
